Add test utilities for counting compilations.
Also uses the new utilities to check that pmap doesn't compile constant computations.
